requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
beautifulsoup4==4.11.1; python_version >= '3.6'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
cryptography==36.0.2
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-allauth==0.50.0
django-bootstrap5==21.3
django-extensions==3.1.5
django==4.0.4
idna==3.3; python_version >= '3'
oauthlib==3.2.0; python_version >= '3.6'
pycparser==2.21
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python3-openid==3.2.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
soupsieve==2.3.2.post1; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
